Bringing camper back on github
Had to transfer the account holding the repository, didnt manage to keep the commits but the content is here and has been cleaned of any old passwords or keys.
This was intended to be an archive of camper and a place for someone to start if they are interested in bringing it back for some use.